chore: replace underscore placeholder with double underscore
Bitcoin uses underscore in util/translation.h for translatable strings
but underscores are also a placeholder used in multiple languages, incl.
C++. The next commit will be introducing backports, one of them will be
placing util/translation.h into validation.h, which is a header used by
Dash-specific code.

This causes a conflict between the normal usage of underscore as a
placeholder and Bitcoin's usage of it as a function name, which is
reported by the Dash-specific linter.

Neither C++17 nor C++20 offer dedicated placeholders (the closest being
std::ignore, which only applies to certain usage of tuples, not
structured bindings) nor do we want to silence the linter when it comes
to reporting potential shadowing.

So we replace our underscore usage for throwaway values with double
underscore and hope no future backport uses that up as well.