rpc: partially allow param conversions for composite commands
The processing is done in RPCConsole::RPCParseCommandLine() instead of
CRPCTable::execute() because we need to be able to populate
vRPCConvertParams with arg positions without having to worry about the
offset-by-one as the parsing is done *before* execute() gets a chance
to account for the subcommand.

Named values are currently still not supported.