net: rename `Peer::m_block_relay_only` to `Peer::m_can_tx_relay`
The true purpose of `m_block_relay_only` is to use it in place of
`m_tx_relay == nullptr`, used to indicate if the node is permitted
to relay transactions.

In upcoming commits, being a block-relay-only node will not be the only
reason to not relay transactions, so let's rename the variable to what
we actually use it for.