ci: upload built binaries as artifacts
Currently, we upload sort-of the contents of what we would expect in
build-ci, meant for consumption by the functional test runner, except
it's too incomplete for the runner to work with it. Plus, we exclude
files that are actually valuable for humans who might value them.

So we'll deal with functional tests in the next commit, this commit is
to ensure humans get binaries they can examine. Retention period's short
though because a run can produce, in total, multi-gigabyte artifacts
(after compression applied on individual artifacts) and we have a storage
budget to abide by.