ci: tentatively drop multiprocess and tsan functional tests
Currently, we are experiencing issues surrounding exhausting disk space
during the log collection of these runs. While actions exist to free up
space on the runner, they are unusable when the job runs in a container
context nor are deleting files a viable way to get by, we will drop this
for now until a remedy for this can be figured out.