evo: require all primary addresses to always match CORE_P2P's address
This is currently done to avoid breaking the current assumption that
Core, Platform and Platform HTTPS API are all hosted on the same address
and only differ in port.

In later versions, we should consider relaxing this requirement but
keeping the restriction vastly simplifies compatibility code and should
make the transition easier.